requirements.txt
Flask==3.0.3
pillow==10.3.0
torch==2.4.1
torchvision==0.19.1
